backend/django/requirements.txt
asgiref
autopep8
Django
django-cors-headers
djangorestframework
djangorestframework-jwt
drf-yasg
djangorestframework-simplejwt
mysqlclient
pycodestyle
PyJWT
pytz
sqlparse
toml
Pillow
requests
